refactor core internals
