+++ .changeset/clean-core-refactor.md
@@ -1,0 +1,5 @@
+---
+"@anvia/core": patch
+---
+
+Refactor core internals for improved maintainability while preserving public API and behavior.
