+++ packages/core/src/agent/request.ts
@@ -3,42 +3,37 @@ import {
   type CompletionModel,
   CompletionRequestBuilder,
   type CompletionResponse,
-  type Document,
   type JsonObject,
   Message,
   type Message as MessageType,
   type ReasoningContentType,
   type ToolCall,
-  ToolContent,
   type ToolDefinition,
   type ToolResult,
   type ToolResultContent,
   textFromAssistantContent,
   Usage,
 } from "../completion/index";
-import type { MemoryContext, MemoryRegistration, MemorySavePolicy } from "../memory";
-import {
-  type ActiveAgentRunObservers,
-  type ActiveToolObservers,
-  startAgentRunObservers,
-} from "../observability/group";
+import { createAsyncQueue } from "../internal/async-queue";
+import type { MemoryContext } from "../memory";
+import { type ActiveAgentRunObservers, startAgentRunObservers } from "../observability/group";
 import type { AgentTraceInfo, AgentTraceOptions } from "../observability/types";
 import { toReadableStream } from "../streaming";
-import {
-  type AnyTool,
-  type NormalizedToolOutput,
-  type ToolCallStreamEvent,
-  toolResultContentToText,
-} from "../tool";
-import type { ToolMiddleware, ToolResultMiddlewareArgs } from "../tool/middleware";
+import type { ToolMiddleware } from "../tool/middleware";
 import type { Agent } from "./agent";
 import { MaxTurnsError, PromptCancelledError } from "./errors";
-import type { PromptHook, ToolHookArgs } from "./hooks";
-import { runControl, toolCallControl } from "./hooks";
+import type { PromptHook } from "./hooks";
+import { runControl } from "./hooks";
+import { PromptRequestMemory } from "./request-memory";
+import { fetchDynamicContext, fetchToolDefinitions } from "./retrieval";
 import { type AgentDeltaEvent, CompletionStreamAccumulator } from "./stream-accumulator";
-import { extractRagText, isStreamingCompletionModel, mapWithConcurrency } from "./utils";
-
-const MCP_TOOL_METADATA_KEY = Symbol.for("anvia.mcp.tool.metadata");
+import {
+  type AgentToolEventPayload,
+  ToolCallExecutor,
+  type ToolExecutionEventPayload,
+  type ToolResultEventPayload,
+} from "./tool-execution";
+import { extractRagText, isStreamingCompletionModel } from "./utils";
 
 export type PromptResponse = {
   output: string;
@@ -120,16 +115,18 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
   private concurrency = 1;
   private traceOptions: AgentTraceOptions | undefined;
   private requestToolMiddlewares: ToolMiddleware[] = [];
+  private readonly memoryRecorder: PromptRequestMemory;
 
   private constructor(
     private readonly agent: Agent<M>,
     private readonly promptMessage: MessageType,
-    private readonly initialHistory: MessageType[] = [],
-    private readonly memoryContext: MemoryContext | undefined = undefined,
+    initialHistory: MessageType[] = [],
+    memoryContext: MemoryContext | undefined = undefined,
   ) {
     this.chatHistory = initialHistory;
     this.maxTurnCount = agent.defaultMaxTurns ?? 0;
     this.activeHook = agent.hook;
+    this.memoryRecorder = new PromptRequestMemory(agent, memoryContext, initialHistory);
   }
 
   static fromAgent<M extends CompletionModel>(
@@ -174,8 +171,8 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
   async send(): Promise<PromptResponse> {
     const runId = globalThis.crypto.randomUUID();
     const newMessages: MessageType[] = [this.promptMessage];
-    await this.prepareMemoryRun(runId, newMessages);
-    const pendingTurnMessages = this.memoryPolicy() === "turn" ? [...newMessages] : [];
+    this.chatHistory = await this.memoryRecorder.prepareRun(runId, newMessages);
+    const pendingTurnMessages = this.memoryRecorder.pendingTurnMessages(newMessages);
     let usage = Usage.empty();
     let currentTurns = 0;
     let lastPrompt = this.promptMessage;
@@ -195,8 +192,8 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
         await this.runCompletionCallHook(prompt, historyForRequest, newMessages);
 
         const ragText = extractRagText(prompt);
-        const dynamicContext = await this.fetchDynamicContext(ragText);
-        const toolDefs = await this.fetchToolDefinitions(ragText);
+        const dynamicContext = await fetchDynamicContext(this.agent, ragText);
+        const toolDefs = await fetchToolDefinitions(this.agent, ragText);
         const request = new CompletionRequestBuilder(this.agent.model, prompt)
           .instructions(this.agent.instructions)
           .messages(historyForRequest)
@@ -215,7 +212,7 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
 
         const assistantMessage = Message.assistant(response.choice, response.messageId);
         newMessages.push(assistantMessage);
-        await this.commitMemoryMessages(
+        await this.memoryRecorder.commitMessages(
           runId,
           currentTurns,
           [assistantMessage],
@@ -225,7 +222,7 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
           (item): item is ToolCall => item.type === "tool_call",
         );
         if (toolCalls.length === 0) {
-          await this.commitCompletedMemoryRun(
+          await this.memoryRecorder.commitCompletedRun(
             runId,
             currentTurns,
             newMessages,
@@ -254,14 +251,19 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
         );
         const toolMessage = Message.tool(toolResults);
         newMessages.push(toolMessage);
-        await this.commitMemoryMessages(runId, currentTurns, [toolMessage], pendingTurnMessages);
-        await this.commitCompletedMemoryTurn(runId, currentTurns, pendingTurnMessages);
+        await this.memoryRecorder.commitMessages(
+          runId,
+          currentTurns,
+          [toolMessage],
+          pendingTurnMessages,
+        );
+        await this.memoryRecorder.commitCompletedTurn(runId, currentTurns, pendingTurnMessages);
       }
 
       throw new MaxTurnsError(this.maxTurnCount, [...this.chatHistory, ...newMessages], lastPrompt);
     } catch (error) {
       await runObservers.error({ error, usage, messages: [...newMessages] });
-      await this.recordMemoryError(runId, error, newMessages);
+      await this.memoryRecorder.recordError(runId, error, newMessages);
       throw error;
     }
   }
@@ -273,8 +275,8 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
 
     const runId = globalThis.crypto.randomUUID();
     const newMessages: MessageType[] = [this.promptMessage];
-    await this.prepareMemoryRun(runId, newMessages);
-    const pendingTurnMessages = this.memoryPolicy() === "turn" ? [...newMessages] : [];
+    this.chatHistory = await this.memoryRecorder.prepareRun(runId, newMessages);
+    const pendingTurnMessages = this.memoryRecorder.pendingTurnMessages(newMessages);
     let usage = Usage.empty();
     let currentTurns = 0;
     let lastPrompt = this.promptMessage;
@@ -304,8 +306,8 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
         await this.runCompletionCallHook(prompt, historyForRequest, newMessages);
 
         const ragText = extractRagText(prompt);
-        const dynamicContext = await this.fetchDynamicContext(ragText);
-        const toolDefs = await this.fetchToolDefinitions(ragText);
+        const dynamicContext = await fetchDynamicContext(this.agent, ragText);
+        const toolDefs = await fetchToolDefinitions(this.agent, ragText);
         const request = new CompletionRequestBuilder(this.agent.model, prompt)
           .instructions(this.agent.instructions)
           .messages(historyForRequest)
@@ -362,7 +364,7 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
 
         const assistantMessage = Message.assistant(response.choice, response.messageId);
         newMessages.push(assistantMessage);
-        await this.commitMemoryMessages(
+        await this.memoryRecorder.commitMessages(
           runId,
           currentTurns,
           [assistantMessage],
@@ -378,7 +380,7 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
 
         if (toolCalls.length === 0) {
           const output = textFromAssistantContent(response.choice);
-          await this.commitCompletedMemoryRun(
+          await this.memoryRecorder.commitCompletedRun(
             runId,
             currentTurns,
             newMessages,
@@ -422,14 +424,19 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
         const toolResults = await toolResultsPromise;
         const toolMessage = Message.tool(toolResults);
         newMessages.push(toolMessage);
-        await this.commitMemoryMessages(runId, currentTurns, [toolMessage], pendingTurnMessages);
-        await this.commitCompletedMemoryTurn(runId, currentTurns, pendingTurnMessages);
+        await this.memoryRecorder.commitMessages(
+          runId,
+          currentTurns,
+          [toolMessage],
+          pendingTurnMessages,
+        );
+        await this.memoryRecorder.commitCompletedTurn(runId, currentTurns, pendingTurnMessages);
       }
 
       throw new MaxTurnsError(this.maxTurnCount, [...this.chatHistory, ...newMessages], lastPrompt);
     } catch (error) {
       await runObservers.error({ error, usage, messages: [...newMessages] });
-      await this.recordMemoryError(runId, error, newMessages);
+      await this.memoryRecorder.recordError(runId, error, newMessages);
       yield await emit({ type: "error", error });
       throw error;
     }
@@ -479,23 +486,6 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
     }
   }
 
-  private toolTraceMetadata(tool: AnyTool | undefined): JsonObject | undefined {
-    if (tool === undefined) {
-      return undefined;
-    }
-    const metadata = (tool as { [MCP_TOOL_METADATA_KEY]?: unknown })[MCP_TOOL_METADATA_KEY];
-    const mcpMetadata =
-      typeof metadata === "object" && metadata !== null
-        ? (metadata as { serverName?: unknown })
-        : undefined;
-    return {
-      approvalRequired: tool.approval !== undefined,
-      ...(typeof mcpMetadata?.serverName === "string" && mcpMetadata.serverName.length > 0
-        ? { mcpServerName: mcpMetadata.serverName }
-        : {}),
-    };
-  }
-
   private async executeToolCalls(
     toolCalls: ToolCall[],
     newMessages: MessageType[],
@@ -507,162 +497,14 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
       toolDefinitions?: ToolDefinition[];
     },
   ): Promise<ToolResult[]> {
-    return mapWithConcurrency(toolCalls, this.concurrency, async (toolCall) => {
-      const args = JSON.stringify(toolCall.function.arguments ?? {});
-      const internalCallId = globalThis.crypto.randomUUID();
-      const hookArgs: ToolHookArgs = {
-        toolName: toolCall.function.name,
-        internalCallId,
-        args,
-      };
-      if (toolCall.callId !== undefined) {
-        hookArgs.toolCallId = toolCall.callId;
-      }
-      const tool = this.agent.getTool(toolCall.function.name);
-      const toolDefinition = observation?.toolDefinitions?.find(
-        (definition) => definition.name === toolCall.function.name,
-      );
-      const toolMetadata = this.toolTraceMetadata(tool);
-
-      const toolObservers = await observation?.runObservers.startTool({
-        turn: observation.turn,
-        toolCall,
-        toolName: toolCall.function.name,
-        internalCallId,
-        args,
-        toolCallId: toolCall.callId,
-        ...(toolDefinition === undefined ? {} : { toolDefinition }),
-        ...(toolMetadata === undefined ? {} : { toolMetadata }),
-      });
-
-      const callAction = await this.activeHook?.onToolCall?.({
-        ...hookArgs,
-        tool: toolCallControl,
-      });
-      if (callAction?.type === "terminate") {
-        await this.recordToolError(
-          toolObservers,
-          observation?.turn,
-          toolCall,
-          internalCallId,
-          args,
-          callAction.reason,
-        );
-        throw this.cancelled(newMessages, callAction.reason);
-      }
-      if (callAction?.type === "approval_request") {
-        const reason = `Tool approval was requested for ${toolCall.function.name}, but no approval handler is installed.`;
-        await this.recordToolError(
-          toolObservers,
-          observation?.turn,
-          toolCall,
-          internalCallId,
-          args,
-          reason,
-        );
-        throw this.cancelled(newMessages, reason);
-      }
-
-      let output: NormalizedToolOutput;
-      let skipped = false;
-      if (callAction?.type === "skip") {
-        output = callAction.reason;
-        skipped = true;
-      } else {
-        try {
-          output = await this.agent.callTool(toolCall.function.name, args, {
-            emitStreamEvent: async (event) => {
-              await toolObservers?.streamEvent({
-                turn: observation?.turn ?? 0,
-                toolCall,
-                toolName: toolCall.function.name,
-                internalCallId,
-                args,
-                ...(toolCall.callId === undefined ? {} : { toolCallId: toolCall.callId }),
-                event,
-              });
-              const payload = agentToolEventPayload(toolCall, internalCallId, event);
-              if (payload !== undefined) {
-                onStreamEvent?.(payload);
-              }
-            },
-          });
-        } catch (error) {
-          output = error instanceof Error ? error.toString() : String(error);
-        }
-      }
-
-      let result = toolOutputToText(output);
-      let structuredResult = toolOutputToStructuredResult(output);
-      if (this.agent.shouldApplyToolMiddleware(toolCall.function.name)) {
-        const middlewareReplacement = await this.runToolResultMiddlewares({
-          ...hookArgs,
-          result,
-          originalResult: result,
-          structuredResult,
-          originalStructuredResult: structuredResult,
-          turn: observation?.turn ?? 0,
-        });
-        if (middlewareReplacement !== undefined) {
-          output = middlewareReplacement;
-          result = middlewareReplacement;
-          structuredResult = undefined;
-        }
-      }
-
-      const resultAction = await this.activeHook?.onToolResult?.({
-        ...hookArgs,
-        result,
-        structuredResult,
-        run: runControl,
-      });
-      await toolObservers?.end({
-        turn: observation?.turn ?? 0,
-        toolCall,
-        toolName: toolCall.function.name,
-        internalCallId,
-        args,
-        result,
-        structuredResult,
-        skipped,
-        toolCallId: toolCall.callId,
-      });
-      if (resultAction?.type === "terminate") {
-        throw this.cancelled(newMessages, resultAction.reason);
-      }
-
-      const resultPayload: ToolResultEventPayload = {
-        type: "tool_result",
-        toolName: toolCall.function.name,
-        internalCallId,
-        args,
-        result,
-        structuredResult,
-      };
-      if (toolCall.callId !== undefined) {
-        resultPayload.toolCallId = toolCall.callId;
-      }
-      onResult?.(resultPayload);
-      return ToolContent.toolResult(toolCall.id, output, toolCall.callId);
-    });
-  }
-
-  private async runToolResultMiddlewares(
-    args: ToolResultMiddlewareArgs,
-  ): Promise<string | undefined> {
-    let result = args.result;
-    let replaced = false;
-    for (const middleware of [...this.agent.toolMiddlewares, ...this.requestToolMiddlewares]) {
-      const replacement = await middleware.onResult?.({
-        ...args,
-        result,
-      });
-      if (replacement !== undefined) {
-        result = replacement;
-        replaced = true;
-      }
-    }
-    return replaced ? result : undefined;
+    const executor = new ToolCallExecutor(
+      this.agent,
+      this.activeHook,
+      this.concurrency,
+      this.requestToolMiddlewares,
+      (reason) => this.cancelled(newMessages, reason),
+    );
+    return executor.execute(toolCalls, onResult, onStreamEvent, observation);
   }
 
   private async startRunObservers(): Promise<ActiveAgentRunObservers> {
@@ -712,84 +554,6 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
     });
   }
 
-  private async fetchDynamicContext(ragText: string | undefined): Promise<Document[]> {
-    if (ragText === undefined || ragText.length === 0 || this.agent.dynamicContexts.length === 0) {
-      return [];
-    }
-
-    const documents: Document[] = [];
-    for (const registration of this.agent.dynamicContexts) {
-      const results = await registration.index.search({
-        query: ragText,
-        topK: registration.options.topK,
-        threshold: registration.options.threshold,
-        filter: registration.options.filter,
-      });
-      for (const result of results) {
-        const formatted = registration.options.format?.(result);
-        if (formatted !== undefined) {
-          documents.push(formatted);
-        } else {
-          const metadata = formatMetadata(result.metadata);
-          documents.push({
-            id: result.id,
-            text:
-              typeof result.document === "string"
-                ? result.document
-                : JSON.stringify(result.document, null, 2),
-            ...(metadata === undefined ? {} : { additionalProps: metadata }),
-          });
-        }
-      }
-    }
-    return documents;
-  }
-
-  private async fetchToolDefinitions(ragText: string | undefined): Promise<ToolDefinition[]> {
-    const staticDefinitions = await this.agent.toolSet.getToolDefinitions(ragText);
-    if (ragText === undefined || ragText.length === 0 || this.agent.dynamicTools.length === 0) {
-      return staticDefinitions;
-    }
-
-    const definitions = [...staticDefinitions];
-    const names = new Set(staticDefinitions.map((definition) => definition.name));
-    for (const registration of this.agent.dynamicTools) {
-      const results = await registration.index.search({
-        query: ragText,
-        topK: registration.options.topK,
-        threshold: registration.options.threshold,
-        filter: registration.options.filter,
-      });
-      for (const result of results) {
-        if (names.has(result.document.toolName)) {
-          continue;
-        }
-        names.add(result.document.toolName);
-        definitions.push(result.document.definition);
-      }
-    }
-    return definitions;
-  }
-
-  private async recordToolError(
-    toolObservers: ActiveToolObservers | undefined,
-    turn: number | undefined,
-    toolCall: ToolCall,
-    internalCallId: string,
-    args: string,
-    error: unknown,
-  ): Promise<void> {
-    await toolObservers?.error({
-      turn: turn ?? 0,
-      toolCall,
-      toolName: toolCall.function.name,
-      internalCallId,
-      args,
-      error,
-      toolCallId: toolCall.callId,
-    });
-  }
-
   private async runCompletionCallHook(
     prompt: MessageType,
     history: MessageType[],
@@ -825,118 +589,6 @@ export class PromptRequest<M extends CompletionModel = CompletionModel> {
   private cancelled(newMessages: MessageType[], reason: string): PromptCancelledError {
     return new PromptCancelledError([...this.chatHistory, ...newMessages], reason);
   }
-
-  private memory(): MemoryRegistration | undefined {
-    return this.memoryContext === undefined ? undefined : this.agent.memory;
-  }
-
-  private memoryPolicy(): MemorySavePolicy | undefined {
-    return this.memory()?.options.savePolicy;
-  }
-
-  private async prepareMemoryRun(runId: string, newMessages: MessageType[]): Promise<void> {
-    const memory = this.memory();
-    if (memory === undefined || this.memoryContext === undefined) {
-      this.chatHistory = this.initialHistory;
-      return;
-    }
-
-    const memoryHistory = await memory.store.load(this.memoryContext);
-    this.chatHistory = [...memoryHistory, ...this.initialHistory];
-    if (memory.options.savePolicy === "message") {
-      await memory.store.append({
-        context: this.memoryContext,
-        runId,
-        turn: 1,
-        messages: newMessages,
-      });
-    }
-  }
-
-  private async commitMemoryMessages(
-    runId: string,
-    turn: number,
-    messages: MessageType[],
-    pendingTurnMessages: MessageType[],
-  ): Promise<void> {
-    const memory = this.memory();
-    if (memory === undefined || this.memoryContext === undefined || messages.length === 0) {
-      return;
-    }
-    if (memory.options.savePolicy === "message") {
-      await memory.store.append({
-        context: this.memoryContext,
-        runId,
-        turn,
-        messages,
-      });
-    } else if (memory.options.savePolicy === "turn") {
-      pendingTurnMessages.push(...messages);
-    }
-  }
-
-  private async commitCompletedMemoryTurn(
-    runId: string,
-    turn: number,
-    pendingTurnMessages: MessageType[],
-  ): Promise<void> {
-    const memory = this.memory();
-    if (
-      memory === undefined ||
-      this.memoryContext === undefined ||
-      memory.options.savePolicy !== "turn" ||
-      pendingTurnMessages.length === 0
-    ) {
-      return;
-    }
-    await memory.store.append({
-      context: this.memoryContext,
-      runId,
-      turn,
-      messages: [...pendingTurnMessages],
-    });
-    pendingTurnMessages.length = 0;
-  }
-
-  private async commitCompletedMemoryRun(
-    runId: string,
-    turn: number,
-    newMessages: MessageType[],
-    pendingTurnMessages: MessageType[],
-  ): Promise<void> {
-    await this.commitCompletedMemoryTurn(runId, turn, pendingTurnMessages);
-    const memory = this.memory();
-    if (
-      memory === undefined ||
-      this.memoryContext === undefined ||
-      memory.options.savePolicy !== "run"
-    ) {
-      return;
-    }
-    await memory.store.append({
-      context: this.memoryContext,
-      runId,
-      turn,
-      messages: [...newMessages],
-    });
-  }
-
-  private async recordMemoryError(
-    runId: string,
-    error: unknown,
-    newMessages: MessageType[],
-  ): Promise<void> {
-    const memory = this.memory();
-    if (memory === undefined || this.memoryContext === undefined) {
-      return;
-    }
-    await memory.store.recordError?.({
-      context: this.memoryContext,
-      runId,
-      error,
-      messages: [...newMessages],
-    });
-  }
 }
 
 function normalizePromptInput(prompt: string | MessageType | MessageType[]): {
@@ -962,137 +614,6 @@ function normalizePromptInput(prompt: string | MessageType | MessageType[]): {
   };
 }
 
-type ToolResultEventPayload = {
-  type: "tool_result";
-  toolName: string;
-  toolCallId?: string;
-  internalCallId: string;
-  args: string;
-  result: string;
-  structuredResult?: ToolResultContent[] | undefined;
-};
-
-type AgentToolEventPayload = {
-  type: "agent_tool_event";
-  toolName: string;
-  toolCallId?: string;
-  internalCallId: string;
-  agentId: string;
-  agentName?: string;
-  event: AgentChildStreamEvent;
-};
-
-type ToolExecutionEventPayload = ToolResultEventPayload | AgentToolEventPayload;
-
-function toolOutputToText(output: NormalizedToolOutput): string {
-  return typeof output === "string" ? output : toolResultContentToText(output);
-}
-
-function toolOutputToStructuredResult(
-  output: NormalizedToolOutput,
-): ToolResultContent[] | undefined {
-  return typeof output === "string" ? undefined : output;
-}
-
-function agentToolEventPayload(
-  toolCall: ToolCall,
-  internalCallId: string,
-  event: ToolCallStreamEvent,
-): AgentToolEventPayload | undefined {
-  if (typeof event.agentId !== "string" || event.agentId.length === 0) {
-    return undefined;
-  }
-  return {
-    type: "agent_tool_event",
-    toolName: toolCall.function.name,
-    ...(toolCall.callId === undefined ? {} : { toolCallId: toolCall.callId }),
-    internalCallId,
-    agentId: event.agentId,
-    ...(event.agentName === undefined ? {} : { agentName: event.agentName }),
-    event: event.event as AgentChildStreamEvent,
-  };
-}
-
-type AsyncQueueWaiter<T> = {
-  resolve: (result: IteratorResult<T>) => void;
-  reject: (error: unknown) => void;
-};
-
-function createAsyncQueue<T>(): AsyncIterable<T> & {
-  enqueue(value: T): void;
-  close(): void;
-  throw(error: unknown): void;
-} {
-  const values: T[] = [];
-  const waiters: AsyncQueueWaiter<T>[] = [];
-  let closed = false;
-  let error: unknown;
-
-  function flush(): void {
-    while (waiters.length > 0 && values.length > 0) {
-      const waiter = waiters.shift();
-      const value = values.shift() as T;
-      if (waiter !== undefined) {
-        waiter.resolve({ value, done: false });
-      }
-    }
-
-    if (values.length > 0 || waiters.length === 0 || !closed) {
-      return;
-    }
-
-    while (waiters.length > 0) {
-      const waiter = waiters.shift();
-      if (waiter === undefined) {
-        continue;
-      }
-      if (error !== undefined) {
-        waiter.reject(error);
-      } else {
-        waiter.resolve({ value: undefined, done: true });
-      }
-    }
-  }
-
-  return {
-    enqueue(value: T): void {
-      if (closed) {
-        return;
-      }
-      values.push(value);
-      flush();
-    },
-    close(): void {
-      closed = true;
-      flush();
-    },
-    throw(thrown: unknown): void {
-      closed = true;
-      error = thrown;
-      flush();
-    },
-    [Symbol.asyncIterator](): AsyncIterator<T> {
-      return {
-        next(): Promise<IteratorResult<T>> {
-          if (values.length > 0) {
-            const value = values.shift() as T;
-            return Promise.resolve({ value, done: false });
-          }
-          if (error !== undefined) {
-            return Promise.reject(error);
-          }
-          if (closed) {
-            return Promise.resolve({ value: undefined, done: true });
-          }
-          return new Promise((resolve, reject) => {
-            waiters.push({ resolve, reject });
-          });
-        },
-      };
-    },
-  };
-}
-
 function addTurn(turn: number, event: AgentDeltaEvent): AgentStreamEvent {
   if (event.type === "text_delta") {
     return { type: "text_delta", turn, delta: event.delta };
@@ -1114,14 +635,4 @@ function isGenerationDeltaEvent(type: string): boolean {
     type === "tool_call_delta" ||
     type === "tool_call"
   );
-}
-
-function formatMetadata(
-  metadata: Record<string, unknown> | undefined,
-): Record<string, string> | undefined {
-  if (metadata === undefined) {
-    return undefined;
-  }
-
-  return Object.fromEntries(Object.entries(metadata).map(([key, value]) => [key, String(value)]));
 }
